chore: add-more-sample-queries
